Remove sharedTest module and move all fake implementations alongside the actual implementation
