+++ ground/build.gradle
@@ -166,8 +166,6 @@ android {
 dependencies {
     implementation fileTree(include: ['*.jar'], dir: 'libs')
     implementation project(':android-gmaps-addons')
-    testImplementation project(':sharedTest')
-    androidTestImplementation project(':sharedTest')
 
     implementation 'androidx.multidex:multidex:2.0.1'
     implementation 'androidx.preference:preference:1.1.1'
@@ -225,6 +223,7 @@ dependencies {
     // ReactiveX
     implementation 'io.reactivex.rxjava2:rxjava:2.2.21'
     implementation "io.reactivex.rxjava2:rxandroid:2.1.1"
+    implementation 'com.squareup.rx.idler:rx2-idler:0.11.0'
     implementation "com.jakewharton.rxbinding2:rxbinding:$project.rxBindingVersion"
     implementation "com.jakewharton.rxbinding2:rxbinding-support-v4:$project.rxBindingVersion"
     implementation "com.jakewharton.rxbinding2:rxbinding-appcompat-v7:$project.rxBindingVersion"
@@ -274,7 +273,9 @@ dependencies {
     implementation 'commons-io:commons-io:2.8.0'
 
     // Testing
+    implementation 'junit:junit:4.13.2'
     testImplementation 'junit:junit:4.13.2'
+    androidTestImplementation 'junit:junit:4.13.2'
     testImplementation 'com.google.truth:truth:1.1.3'
     androidTestImplementation 'com.google.truth:truth:1.1.3'
     testImplementation 'androidx.test:core:1.3.0'
@@ -289,15 +290,15 @@ dependencies {
     androidTestImplementation 'org.mockito:mockito-core:4.4.0'
     testImplementation "org.mockito.kotlin:mockito-kotlin:4.0.0"
 
-    androidTestImplementation('androidx.test.espresso:espresso-core:3.3.0', {
-        exclude group: 'com.android.support', module: 'support-annotations'
-    })
-    androidTestImplementation 'junit:junit:4.13.2'
     androidTestImplementation 'androidx.test.ext:junit:1.1.2'
     androidTestImplementation 'android.arch.core:core-testing:1.1.1'
-    androidTestImplementation 'com.squareup.rx.idler:rx2-idler:0.11.0'
-    testImplementation 'com.squareup.rx.idler:rx2-idler:0.11.0'
+    implementation('androidx.test.espresso:espresso-core:3.3.0', {
+        exclude group: 'com.android.support', module: 'support-annotations'
+    })
     testImplementation('androidx.test.espresso:espresso-core:3.3.0', {
+        exclude group: 'com.android.support', module: 'support-annotations'
+    })
+    androidTestImplementation('androidx.test.espresso:espresso-core:3.3.0', {
         exclude group: 'com.android.support', module: 'support-annotations'
     })
