+++ ground/src/androidTest/java/com/google/android/ground/TestWorkManagerModule.java
@@ -17,8 +17,8 @@
 package com.google.android.ground;
 
 import androidx.work.WorkManager;
+import com.google.android.ground.persistence.sync.FakeWorkManager;
 import com.google.android.ground.persistence.sync.WorkManagerModule;
-import com.google.android.ground.test.persistence.sync.FakeWorkManager;
 import dagger.Module;
 import dagger.Provides;
 import dagger.hilt.components.SingletonComponent;
