+++ ground/src/main/java/com/google/android/ground/system/auth/FakeAuthenticationManager.java
@@ -14,14 +14,13 @@
  * limitations under the License.
  */
 
-package com.google.android.ground.test.system.auth;
+package com.google.android.ground.system.auth;
 
+import androidx.annotation.NonNull;
+import com.google.android.ground.FakeData;
 import com.google.android.ground.model.User;
 import com.google.android.ground.rx.annotations.Hot;
-import com.google.android.ground.system.auth.AuthenticationManager;
-import com.google.android.ground.system.auth.SignInState;
 import com.google.android.ground.system.auth.SignInState.State;
-import com.google.android.ground.test.FakeData;
 import io.reactivex.Observable;
 import io.reactivex.subjects.BehaviorSubject;
 import io.reactivex.subjects.Subject;
@@ -40,11 +39,13 @@ public class FakeAuthenticationManager implements AuthenticationManager {
   @Inject
   public FakeAuthenticationManager() {}
 
+  @NonNull
   @Override
   public Observable<SignInState> getSignInState() {
     return behaviourSubject;
   }
 
+  @NonNull
   @Override
   public User getCurrentUser() {
     return user;
