+++ ground/src/test/java/com/google/android/ground/TestObservers.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.google.android.ground.test;
+package com.google.android.ground;
 
 import androidx.lifecycle.LiveData;
 import androidx.lifecycle.Observer;
