+++ ground/src/test/java/com/google/android/ground/TestSchedulersModule.java
@@ -18,7 +18,7 @@ package com.google.android.ground;
 
 import com.google.android.ground.rx.Schedulers;
 import com.google.android.ground.rx.SchedulersModule;
-import com.google.android.ground.test.TestScheduler;
+import com.google.android.ground.rx.TestScheduler;
 import dagger.Binds;
 import dagger.Module;
 import dagger.hilt.components.SingletonComponent;
