+++ ground/src/test/java/com/google/android/ground/repository/LocationOfInterestRepositoryTest.java
@@ -25,6 +25,7 @@ import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
 import com.google.android.ground.BaseHiltTest;
+import com.google.android.ground.FakeData;
 import com.google.android.ground.model.mutation.LocationOfInterestMutation;
 import com.google.android.ground.model.mutation.Mutation;
 import com.google.android.ground.model.mutation.Mutation.SyncStatus;
@@ -32,12 +33,11 @@ import com.google.android.ground.model.mutation.Mutation.Type;
 import com.google.android.ground.persistence.local.LocalDataStore;
 import com.google.android.ground.persistence.local.LocalDataStoreModule;
 import com.google.android.ground.persistence.local.room.models.MutationEntitySyncStatus;
+import com.google.android.ground.persistence.remote.FakeRemoteDataStore;
 import com.google.android.ground.persistence.remote.NotFoundException;
 import com.google.android.ground.persistence.remote.RemoteDataEvent;
 import com.google.android.ground.persistence.sync.DataSyncWorkManager;
-import com.google.android.ground.test.FakeData;
-import com.google.android.ground.test.persistence.remote.FakeRemoteDataStore;
-import com.google.android.ground.test.system.auth.FakeAuthenticationManager;
+import com.google.android.ground.system.auth.FakeAuthenticationManager;
 import com.google.common.collect.ImmutableSet;
 import dagger.hilt.android.testing.BindValue;
 import dagger.hilt.android.testing.HiltAndroidTest;
