+++ ground/src/test/java/com/google/android/ground/ui/home/featuredetails/MoveMenuVisibilityTest.java
@@ -18,9 +18,9 @@ package com.google.android.ground.ui.home.featuredetails;
 
 import static com.google.common.truth.Truth.assertThat;
 
+import com.google.android.ground.TestObservers;
 import com.google.android.ground.model.User;
 import com.google.android.ground.model.locationofinterest.LocationOfInterest;
-import com.google.android.ground.test.TestObservers;
 import dagger.hilt.android.testing.HiltAndroidTest;
 import java.util.Arrays;
 import java.util.Collection;
