+++ ground/src/test/java/com/google/android/ground/ui/home/mapcontainer/PolygonDrawingViewModelTest.java
@@ -20,8 +20,8 @@ import static com.google.common.truth.Truth.assertThat;
 import static org.junit.Assert.assertThrows;
 
 import com.google.android.ground.BaseHiltTest;
+import com.google.android.ground.FakeData;
 import com.google.android.ground.model.locationofinterest.Point;
-import com.google.android.ground.test.FakeData;
 import com.google.android.ground.ui.home.mapcontainer.PolygonDrawingViewModel.PolygonDrawingState;
 import com.google.android.ground.ui.map.MapLocationOfInterest;
 import com.google.android.ground.ui.map.MapPin;
