+++ settings.gradle
@@ -1,3 +1,3 @@
-include ':ground', ':android-gmaps-addons', ':sharedTest'
+include ':ground', ':android-gmaps-addons'
 
 project(':android-gmaps-addons').projectDir = new File('third_party/android-gmaps-addons')
