--- sharedTest/build.gradle
@@ -1,59 +1,0 @@
-/*
- * Copyright 2022 Google LLC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-apply plugin: 'com.android.library'
-apply plugin: 'kotlin-android'
-
-apply from: '../config/checkstyle/checkstyle.gradle'
-apply from: '../config/lint/lint.gradle'
-apply from: '../config/pmd/pmd.gradle'
-apply from: '../config/spotbugs/spotbugs.gradle'
-apply from: '../config/jacoco/jacoco.gradle'
-
-android {
-    namespace 'com.google.android.ground.test'
-    compileSdkVersion rootProject.androidCompileSdk
-
-    defaultConfig {
-        minSdkVersion rootProject.androidMinSdk
-        targetSdkVersion rootProject.androidTargetSdk
-    }
-
-    buildTypes {
-        staging {
-        }
-    }
-}
-
-dependencies {
-    implementation project(':ground')
-
-    // Room
-    implementation 'androidx.room:room-runtime:2.4.3'
-
-    // Guava
-    implementation 'com.google.guava:guava:30.1-jre'
-
-    // Support for Java streams.
-    implementation 'net.sourceforge.streamsupport:streamsupport-cfuture:1.7.4'
-
-    // Hilt
-    implementation 'androidx.hilt:hilt-work:1.0.0'
-
-    // ReactiveX
-    implementation 'io.reactivex.rxjava2:rxjava:2.2.21'
-    implementation 'com.squareup.rx.idler:rx2-idler:0.11.0'
-}
